+++ src/pages/SpecialRequest/specialRequestList.tsx
@@ -225,6 +225,18 @@ export const SpecialRequestList: React.FC = () => {
       },
     },
     {
+      title: "เขต",
+      dataIndex: "customerZone",
+      key: "customerZone",
+      render: (value: any, row: any, index: number) => {
+        return (
+          <FlexCol>
+            <Text level={5}>{value || "-"}</Text>
+          </FlexCol>
+        );
+      },
+    },
+    {
       title: "ทั้งหมด",
       dataIndex: "totalPrice",
       key: "totalPrice",